[Update] cal_nutrients.py & login
Former-commit-id: f0ccf7edb3736de4d2f83ae60c7bdf1ccc31be65

+++ app/views.py
@@ -67,10 +67,10 @@ def diet_food():
 
     if request.method == "POST" and request.form.get('btn') == 'form_personal' :
         if request.form.get('age') and request.form.get('gender') and request.form.get('height') and request.form.get('activity') :
-            age = int(request.form.get('age'))
+            age = float(request.form.get('age'))
             sex = request.form.get('gender')
-            height = int(request.form.get('height'))
-            activity = int(request.form.get('activity'))
+            height = float(request.form.get('height'))
+            activity = float(request.form.get('activity'))
             Z = cal_nutrients.body_classifier(sex, age, height)
             nutrients = cal_nutrients.nutrient(Z, sex, age, activity)
             return redirect(url_for("diet_food"))
